Leaderboard and graph links
I added the leaderboard table with graphs for the teachers dashboard, i also added positions of students to make it more competitive. I added the link between parent to child as well as the parent being able to view the chillds information and data. I have also added the ability for teachers to be able to add parents and remove them from the child accounts.